requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Statistics & forecasting
statsmodels>=0.14.0
prophet>=1.1.5
scikit-learn>=1.3.0
# pmdarima removed — fixed SARIMAX orders used instead of auto_arima

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
langchain-classic>=1.0.0

# Vector store
chromadb>=0.5.0

# Visualization
plotly>=5.18.0

# UI
streamlit>=1.35.0

# Environment
python-dotenv>=1.0.0
